clearProfilePushToken on logout
